Add toString() method to err object
Node calls toString() on error objects that have the method defined. Previously, if a request did not pass validation, the response body was `[object Object]`. Now the toString() method will be called, which stringifies and returns err.errors (an array of validation errors). This stringified JSON will now be the body of the response.